Set PATH and exec-path from shell $PATH
On OS X, it's rare to launch Emacs from a process that has inherited the same
environment as your login shell. I was running into the case that I wouldn't
be able to save a golang buffer until I had launched emacs with my shell, so
that goimports would be on my PATH.

To fix this, setting up the path in emacs correctly from the shell seems to
be the optimal solution. The alternative that I had originally considered was
to just add $HOME/go/bin (or $GOPATH/bin) to my exec-path, but this feels like
a robust pattern (for OS X).